adding app configuration
