+++ src/index.ts
@@ -1,21 +1,3 @@
-import express from 'express';
-import dotenv from 'dotenv';
-import cors from 'cors';
+import {createApp} from "./app";
 
-dotenv.config();
-
-const app = express();
-const port = process.env.PORT || 3000;
-
-app.disable("x-powered-by");
-app.use(express.json());
-app.use(cors());
-
-//const userModel = new UserModel();
-
-// Routes
-//app.use('/api', appRouter(userModel));
-
-app.listen(port, () => {
-    console.log(`Server running on http://localhost:${port}`);
-});
+createApp();
