+++ src/router.ts
@@ -1,0 +1,7 @@
+import {Router} from "express";
+// TODO import {Models} from "./types";
+
+export const appRouter = () => {
+    const router = Router();
+    return router;
+};
